saramin-scrap/requirements.txt
selenium
beautifulsoup4
pandas
python-dotenv
fake-useragent
lxml
webdriver-manager
reportlab
pillow
